Created separate test file for raster module
